refactor: shared component 정리
