+++ app/components/form/TextEditor.tsx
@@ -1,6 +1,6 @@
 import { marked } from 'marked';
 import { useState } from 'react';
-import Tooltip from '~/components/Tooltip';
+import Tooltip from '~/components/shared/Tooltip';
 
 export function TextEditor({
   id,
