+++ app/routes/plan.mdx
@@ -7,7 +7,7 @@ import plan from "../components/plan.css";
 export const links = () => [
   { rel: "stylesheet", href: plan },
 ];
-import RemixLogo from '../components/RemixLogo'
+import RemixLogo from '../components/icons/RemixLogo'
 import CenterCardLayout from '../components/CenterCardLayout'
 
 <CenterCardLayout>
